fixes most comments
