+++ flow-typed/overrides/Image.js
@@ -1,0 +1,4 @@
+ declare class Image extends HTMLImageElement {
+   constructor(width?: number, height?: number): void;
+   referrerPolicy: ReferrerPolicyType;
+ }
