+++ src/common/types/profile.js
@@ -115,13 +115,3 @@ export type Profile = {
   tasktracer: TaskTracer,
   threads: Thread[],
 };
-
-export type Node = {
-  totalTime: string,
-  totalTimePercent: string,
-  selfTime: string,
-  name: string,
-  lib: string,
-  dim: boolean,
-  icon: string | null,
-};
